config/configtls: gate TPM key-file support behind !requirefips build tag
go-tpm-keyfiles uses ChaCha20-Poly1305 for TPM importable-key wrapping,
which is not a FIPS 140-approved algorithm. When building with the
requirefips tag (GOFIPS140=v1.0), this pulled non-FIPS crypto into the
binary unconditionally, even when TPM keys were never configured.

Apply the same pattern used in curves_fips.go / curves_nofips.go:
- tpm_nofips.go (!requirefips): full implementation, unchanged
- tpm_fips.go (requirefips): stub that returns an explicit error; imports
  only go-tpm/tpm2/transport for method-signature compatibility, not
  go-tpm-keyfiles
- tpm_test.go renamed to tpm_nofips_linux_test.go (linux && !requirefips):
  simulator requires OpenSSL headers not available on Windows or macOS
- tpm_fips_test.go (requirefips): verifies the stub returns the expected
  error when TPM is enabled in a FIPS build
- tpm_open_linux_test.go (linux): TestTPM_open tests path-handling only,
  no crypto dependency, runs in both FIPS and non-FIPS builds
- README: document that TPM key loading is unsupported in FIPS mode
- Changelog entry added, go.mod tidied

Resolves #13925
