+++ .chloggen/fips-tpm-gate.yaml
@@ -1,0 +1,25 @@
+# Use this changelog template to create an entry for release notes.
+
+# One of 'breaking', 'deprecation', 'new_component', 'enhancement', 'bug_fix'
+change_type: enhancement
+
+# The name of the component, or a single word describing the area of concern, (e.g. receiver/otlp)
+component: pkg/config/configtls
+
+# A brief description of the change.  Surround your text with quotes ("") if it needs to start with a backtick (`).
+note: Gate TPM key loading behind the `!requirefips` build tag. When building with `requirefips` (FIPS 140-3 mode), configuring a TPM key now returns an explicit error instead of pulling ChaCha20-Poly1305 (a non-FIPS algorithm) into the binary via `go-tpm-keyfiles`.
+
+# One or more tracking issues or pull requests related to the change
+issues: [13925]
+
+# (Optional) One or more lines of additional information to render under the primary note.
+# These lines will be padded with 2 spaces and then inserted directly into the document.
+# Use pipe (|) for multiline entries.
+subtext:
+
+# Optional: The change log or logs in which this entry should be included.
+# e.g. '[user]' or '[user, api]'
+# Include 'user' if the change is relevant to end users.
+# Include 'api' if there is a library API change.
+# Default: '[user]'
+change_logs: [user]
