+++ config/configtls/go.mod
@@ -9,7 +9,6 @@ require (
 	github.com/stretchr/testify v1.11.1
 	go.opentelemetry.io/collector/config/configopaque v1.62.0
 	go.opentelemetry.io/collector/confmap v1.62.0
-	go.opentelemetry.io/collector/internal/testutil v0.156.0
 )
 
 require (
@@ -39,4 +38,3 @@ replace go.opentelemetry.io/collector/featuregate => ../../featuregate
 
 replace go.opentelemetry.io/collector/confmap => ../../confmap
 
-replace go.opentelemetry.io/collector/internal/testutil => ../../internal/testutil
