+++ config/configtls/tpm_nofips.go
@@ -1,6 +1,8 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+//go:build !requirefips
+
 package configtls // import "go.opentelemetry.io/collector/config/configtls"
 
 import (
