+++ config/configtls/tpm_nofips_linux_test.go
@@ -1,8 +1,8 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-// Don't run this test on Windows, as it requires a TPM simulator which depends on openssl headers.
-//go:build !windows && !darwin
+// The TPM simulator used here depends on OpenSSL headers not available on Windows or macOS.
+//go:build linux && !requirefips
 
 package configtls // import "go.opentelemetry.io/collector/config/configtls"
 
@@ -18,7 +18,6 @@ import (
 	"math/big"
 	"net"
 	"os"
-	"path/filepath"
 	"testing"
 	"time"
 
@@ -28,12 +27,9 @@ import (
 	"github.com/google/go-tpm/tpm2/transport/simulator"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
-	"go.opentelemetry.io/collector/internal/testutil"
 )
 
 func TestTPM_loadCertificate(t *testing.T) {
-	testutil.SkipIfFIPSOnly(t, "use of CFB is not allowed in FIPS 140-only mode")
 	tpm, err := simulator.OpenSimulator()
 	require.NoError(t, err)
 	defer tpm.Close()
@@ -101,7 +97,6 @@ func TestTPM_loadCertificate_error(t *testing.T) {
 }
 
 func TestTPM_tpmCertificate_errors(t *testing.T) {
-	testutil.SkipIfFIPSOnly(t, "use of CFB is not allowed in FIPS 140-only mode")
 	tpm, err := simulator.OpenSimulator()
 	require.NoError(t, err)
 	defer tpm.Close()
@@ -150,43 +145,6 @@ VGhpcyBpcyBub3QgYSBjZXJ0aWZpY2F0ZS4=
 			tpmCfg := &TPMConfig{}
 			_, err = tpmCfg.tpmCertificate([]byte(test.key), []byte(test.cert), test.openTPM)
 			assert.EqualError(t, err, test.err)
-		})
-	}
-}
-
-func TestTPM_open(t *testing.T) {
-	socketPath := filepath.Join(t.TempDir(), "app.sock")
-	listener, err := net.Listen("unix", socketPath)
-	require.NoError(t, err)
-	defer listener.Close()
-
-	tests := []struct {
-		path string
-		err  string
-	}{
-		{
-			path: "",
-			err:  "TPM path is not set",
-		},
-		{
-			path: "/foo",
-			err:  "failed to open TPM (/foo): stat /foo: no such file or directory",
-		},
-		{
-			path: socketPath,
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.path, func(t *testing.T) {
-			tpm, openErr := openTPM(test.path)()
-			if test.err != "" {
-				require.Nil(t, tpm)
-				assert.Equal(t, test.err, openErr.Error())
-			} else {
-				require.NoError(t, openErr)
-				require.NotNil(t, tpm)
-				tpm.Close()
-			}
 		})
 	}
 }
